AgingData {ospsuite}	R Documentation

AgingData

Description

Results of a sensitivity analysis run (either individual or population simulation)

Super classes

ospsuite::Printable -> ospsuite::DotNetWrapper -> AgingData

Active bindings

individualIds

Array of individual ids (corresponding to the individualId column of the aging table)

parameterPaths

Array of parameter paths (corresponding to the ParameterPath column of the aging table)

times

Array of time values (corresponding to the Time column of the aging table)

values

Array of parameter values (corresponding to the Value column of the aging table)

Methods

Public methods

AgingData$new()

AgingData$print()

Inherited methods

ospsuite::DotNetWrapper$finalize()

Method new()

Initialize a new instance of the class

Usage

AgingData$new()

Returns

A new OSPSuite.R.Domain.AgingData object.

Method print()

Print the object to the console

Usage

AgingData$print(...)

Arguments

...

Rest arguments.

[Package ospsuite version 10.0.70 Index]