Application {ospsuite}	R Documentation

Application

Description

Application representing one instance of an application in a Simulation

Super classes

ospsuite::Printable -> ospsuite::DotNetWrapper -> Application

Active bindings

startTime

Start time of application instance (Read-Only)

infusionTime

Infusion time of application instance or null if undefined (Read-Only)

drugMass

Applied drugmass of the application instance of null if undefined (Read-Only)

Methods

Public methods

Application$print()

Inherited methods

ospsuite::DotNetWrapper$finalize()

ospsuite::DotNetWrapper$initialize()

Method print()

Print the object to the console

Usage

Application$print(...)

Arguments

...

Rest arguments.

[Package ospsuite version 10.0.70 Index]