DataRepository

An object typically holding observed data

Super classes

ospsuite.utils::Printable -> ospsuite::DotNetWrapper -> DataRepository

Active bindings

name

The name of the object.

baseGrid

Returns the base column for the data repository (typically time column).

columns

Returns all columns (including baseGrid) defined in the data repository.

allButBaseGrid

Returns all columns excluding baseGrid defined on the data repository.

metaData

Returns a named list of meta data defined for the data repository. where the name is the name of the metaData and the value is the meta data value.

Methods

Public methods

DataRepository$addColumn()

DataRepository$new()

DataRepository$print()

DataRepository$addMetaData()

DataRepository$removeMetaData()

Method addColumn()

Adds a column to the data repository

Usage

DataRepository$addColumn(column)

Arguments

column

Column to add

Method new()

Initialize a new instance of the class

Usage

DataRepository$new(ref = NULL)

Arguments

ref

Optional underlying DataRepository. If it is not provided, a new instance will be created

Returns

A new DataRepository object.

Method print()

Print the object to the console

Usage

DataRepository$print(...)

Arguments

...

Rest arguments.

Method addMetaData()

Adds a new entry to meta data list or changes its value if the name is already present.

Usage

DataRepository$addMetaData(name, value)

Arguments

name

Name of new meta data list entry

value

Value of new meta data list entry

Method removeMetaData()

Removes the meta data entry in the list if one is defined with this name

Usage

DataRepository$removeMetaData(name)

Arguments

name

Name of meta data entry to delete