MoleculeOntogeny {ospsuite}	R Documentation

MoleculeOntogeny

Description

Use when retrieving individual values using the createIndividualAlgorithm. This class is a simple pair MoleculeName, Ontogeny allowing the user to retrieve potential ontogeny values.

Super class

ospsuite::Printable -> MoleculeOntogeny

Public fields

molecule

Name of the molecule in the model

ontogeny

Name of the ontogeny to use for the molecule

Methods

Public methods

MoleculeOntogeny$new()

MoleculeOntogeny$print()

MoleculeOntogeny$printMoleculeOntogeny()

MoleculeOntogeny$clone()

Inherited methods

Method new()

Initialize a new instance of the class

Usage

MoleculeOntogeny$new(molecule, ontogeny)

Arguments

molecule

molecule name

ontogeny

ontogeny to use for the Molecule (one of StandardOntogeny)

Returns

A new MoleculeOntogeny object.

Method print()

Print the object to the console

Usage

MoleculeOntogeny$print(...)

Arguments

...

Rest arguments.

Method printMoleculeOntogeny()

Print the MoleculeOntogeny on one line

Usage

MoleculeOntogeny$printMoleculeOntogeny()

Method clone()

The objects of this class are cloneable with this method.

Usage

MoleculeOntogeny$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

[Package ospsuite version 10.0.70 Index]