MoleculeParameter {ospsuite}	R Documentation

Standard molecule parameter names typically available in an endogenous molecule (enzyme, transporter etc...) coming from PK-Sim

Description

Standard molecule parameter names typically available in an endogenous molecule (enzyme, transporter etc...) coming from PK-Sim

Usage

MoleculeParameter

Format

An object of class list of length 5.

[Package ospsuite version 10.0.70 Index]