MoleculesBuildingBlock

A Molecules building block. Subclass of BuildingBlock.

Super classes

rSharp::NetObject -> ospsuite::DotNetWrapper -> ospsuite::ObjectBase -> ospsuite::BuildingBlock -> MoleculesBuildingBlock

Methods

Public methods

MoleculesBuildingBlock$new()

MoleculesBuildingBlock$allMoleculeNames()

MoleculesBuildingBlock$allFloatingMoleculeNames()

MoleculesBuildingBlock$allStationaryMoleculeNames()

MoleculesBuildingBlock$allMoleculeNamesOfType()

MoleculesBuildingBlock$allXenobioticFloatingMoleculeNames()

MoleculesBuildingBlock$allEndogenousStationaryMoleculeNames()

MoleculesBuildingBlock$moleculeTypeFor()

Inherited methods

rSharp::NetObject$.printClass()

rSharp::NetObject$.printLine()

rSharp::NetObject$call()

rSharp::NetObject$get()

rSharp::NetObject$getFields()

rSharp::NetObject$getMemberSignature()

rSharp::NetObject$getMethods()

rSharp::NetObject$getProperties()

rSharp::NetObject$getStaticFields()

rSharp::NetObject$getStaticMethods()

rSharp::NetObject$getStaticProperties()

rSharp::NetObject$set()

ospsuite::BuildingBlock$print()

Method new()

Initialize a new instance of the class.

Usage

MoleculesBuildingBlock$new(netObject)

Arguments

netObject

Reference to the underlying Molecules building block.

Returns

A new MoleculesBuildingBlock object.

Method allMoleculeNames()

Returns the names of all molecules defined in the building block.

Usage

MoleculesBuildingBlock$allMoleculeNames()

Returns

Character vector of molecule names.

Method allFloatingMoleculeNames()

Returns the names of all floating molecules in the building block (molecules with IsFloating = TRUE, e.g. drugs, metabolites).

Usage

MoleculesBuildingBlock$allFloatingMoleculeNames()

Returns

Character vector of molecule names.

Method allStationaryMoleculeNames()

Returns the names of all stationary molecules in the building block (molecules with IsFloating = FALSE, e.g. enzymes, transporters).

Usage

MoleculesBuildingBlock$allStationaryMoleculeNames()

Returns

Character vector of molecule names.

Method allMoleculeNamesOfType()

Returns the names of all molecules of the given type. Pass MoleculeType$Protein to obtain all proteins (the union of Enzyme, Transporter, and Binding Partner).

Usage

MoleculesBuildingBlock$allMoleculeNamesOfType(moleculeType)

Arguments

moleculeType

One of the values defined in the MoleculeType enum (e.g. Drug, Metabolite, Enzyme, Transporter, Binding Partner, Complex, Protein).

Returns

Character vector of molecule names.

Method allXenobioticFloatingMoleculeNames()

Returns the names of all xenobiotic floating molecules in the building block (e.g. drugs, drug complexes).

Usage

MoleculesBuildingBlock$allXenobioticFloatingMoleculeNames()

Returns

Character vector of molecule names.

Method allEndogenousStationaryMoleculeNames()

Returns the names of all endogenous stationary molecules in the building block (e.g. enzymes, transporters, other proteins).

Usage

MoleculesBuildingBlock$allEndogenousStationaryMoleculeNames()

Returns

Character vector of molecule names.

Method moleculeTypeFor()

Returns the type of a molecule by name as it is recorded in the building block (e.g. "Drug", "Enzyme", "Transporter", "Binding Partner"). Throws an error if the molecule is not present.

Usage

MoleculesBuildingBlock$moleculeTypeFor(moleculeName)

Arguments

moleculeName

Name of the molecule to look up.

Returns

Character string with the molecule type.