PKParameter

Standard PK Parameters defined in the OSPSuite

Super classes

ospsuite::Printable -> ospsuite::DotNetWrapper -> PKParameter

Active bindings

name

Name of the PK-Parameter

displayName

Display Name of the PK-Parameter. If not set, Name will be used

dimension

Dimension instance used by the PK-Parameter (Read-Only)

unit

Unit of the PK-Parameter (Read-Only)

displayUnit

Display Unit used for the PK-Parameter

Methods

Public methods

PKParameter$print()

PKParameter$clone()

Inherited methods

ospsuite::DotNetWrapper$finalize()

ospsuite::DotNetWrapper$initialize()

Method print()

Print the object to the console

Usage

PKParameter$print(...)

Arguments

...

Rest arguments.

Method print()

Print the object to the console

Usage

PKParameter$print(...)

Arguments

...

Rest arguments.

Method clone()

The objects of this class are cloneable with this method.

Usage

PKParameter$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Method clone()

The objects of this class are cloneable with this method.

Usage

PKParameter$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.