Parameter {ospsuite}	R Documentation

Parameter

Description

A model parameter

Details

Derived from Quantity, please see base class documentation.

Super classes

ospsuite::Printable -> ospsuite::DotNetWrapper -> ospsuite::ObjectBase -> ospsuite::Entity -> ospsuite::Quantity -> Parameter

Active bindings

isStateVariable

Returns TRUE is the parameter has a RHS otherwise FALSE.

rhsFormula

An instance of a Formula object representing the RHS Formula (Read-Only)

Methods

Public methods

Parameter$new()

Parameter$print()

Inherited methods

ospsuite::DotNetWrapper$finalize()

ospsuite::Quantity$hasUnit()

ospsuite::Quantity$printQuantityValue()

ospsuite::Quantity$printValue()

ospsuite::Quantity$reset()

ospsuite::Quantity$setValue()

Method new()

Initialize a new instance of the class

Usage

Parameter$new(ref)

Arguments

ref

.NET Instance

Details

Setting the value to FALSE will delete the RHS Formula. Setting it to TRUE is not currently supported and will throw an error

Returns

A new Parameter object.

Method print()

Print the object to the console

Usage

Parameter$print(...)

Arguments

...

Rest arguments.

[Package ospsuite version 10.0.70 Index]