ParameterRange {ospsuite}	R Documentation

ParameterRange

Description

A parameter range typically used in the definition of PopulationCharacteristics covariates (Height, Weight etc...)

Super classes

ospsuite::Printable -> ospsuite::DotNetWrapper -> ParameterRange

Active bindings

min

Minimum value for the parameter range

max

Maximum value for the parameter range

unit

Unit in which the value is defined

Methods

Public methods

ParameterRange$new()

ParameterRange$print()

ParameterRange$printValue()

ParameterRange$clone()

Inherited methods

ospsuite::DotNetWrapper$finalize()

Method new()

Initialize a new instance of the class

Usage

ParameterRange$new(ref = NULL, min = NULL, max = NULL, unit = NULL)

Arguments

ref

Optional .NET reference object. If not defined, a new instance will be created

min

Optional minimum value for the range

max

Optional minimum value for the range

unit

Optional unit of the specified min and max

Returns

A new ParameterRange object.

Method print()

Print the object to the console

Usage

ParameterRange$print(...)

Arguments

...

Rest arguments.

Method printValue()

Print the the parameter in one line

Usage

ParameterRange$printValue(caption)

Arguments

caption

Caption to display before the value of the parameter

Method clone()

The objects of this class are cloneable with this method.

Usage

ParameterRange$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

[Package ospsuite version 10.0.70 Index]