PopulationCharacteristics {ospsuite}	R Documentation

PopulationCharacteristics

Description

Characteristics of a population used for population creation

Super classes

ospsuite::Printable -> ospsuite::DotNetWrapper -> PopulationCharacteristics

Active bindings

numberOfIndividuals

Number of individuals in the population

proportionOfFemales

Proportion of female in the population

species

Specifies the species of the individual. It should be a species available in PK-Sim (see Species)

population

For a Human species, the population of interest. It should be a population available in PK-Sim (see HumanPopulation)

age

Age range of the population as in instance of a ParameterRange (optional)

gestationalAge

Gestational Age range of the population as in instance of a ParameterRange (optional)

weight

Weight range of the population as in instance of a ParameterRange (optional)

height

Height range of the population as in instance of a ParameterRange (optional)

BMI

BMI range of the population as in instance of a ParameterRange (optional)

allMoleculeOntogenies

All molecule ontogenies defined for this individual characteristics.

seed

Seed used to generate the population

Methods

Public methods

PopulationCharacteristics$new()

PopulationCharacteristics$print()

PopulationCharacteristics$addMoleculeOntogeny()

Inherited methods

ospsuite::DotNetWrapper$finalize()

Method new()

Initialize a new instance of the class

Usage

PopulationCharacteristics$new()

Returns

A new PopulationCharacteristics object.

Method print()

Print the object to the console

Usage

PopulationCharacteristics$print(...)

Arguments

...

Rest arguments.

Method addMoleculeOntogeny()

Add a molecule ontogeny MoleculeOntogeny to the individual characteristics

Usage

PopulationCharacteristics$addMoleculeOntogeny(moleculeOntogeny)

Arguments

moleculeOntogeny

Molecule ontogeny to add

[Package ospsuite version 10.0.70 Index]