QuantityPKParameter {ospsuite}	R Documentation

QuantityPKParameter

Description

pK-Parameter values for all individuals of a simulation (1 or more) calculated for a specific quantity with path quantityPath

Super classes

ospsuite::Printable -> ospsuite::DotNetWrapper -> QuantityPKParameter

Active bindings

values

All values for quantityPath and name

quantityPath

The path of the quantity for which the values were calculated

name

The name of the pK-Parameter (AUC, Cmax, Tmax etc...)

unit

Base unit in which the pk parameter was calculated

dimension

Dimension in which the pk parameter was calculated

Methods

Public methods

QuantityPKParameter$new()

QuantityPKParameter$print()

QuantityPKParameter$clone()

Inherited methods

ospsuite::DotNetWrapper$finalize()

Method new()

Initialize a new instance of the class

Usage

QuantityPKParameter$new(ref)

Arguments

ref

.NET Instance

Returns

A new QuantityPKParameter object.

Method print()

Print the object to the console

Usage

QuantityPKParameter$print(...)

Arguments

...

Rest arguments.

Method clone()

The objects of this class are cloneable with this method.

Usage

QuantityPKParameter$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

[Package ospsuite version 10.0.70 Index]