SimulationBatchOptions {ospsuite}	R Documentation

SimulationBatchOptions

Description

Options to be passed to the SimulationBatch

Super classes

ospsuite::Printable -> ospsuite::DotNetWrapper -> SimulationBatchOptions

Active bindings

variableParameters

Vector of absolute parameter paths to be varied in a simulation batch

variableMolecules

Vector of absolute molecule paths to be varied in a simulation batch

Methods

Public methods

SimulationBatchOptions$new()

SimulationBatchOptions$print()

Inherited methods

ospsuite::DotNetWrapper$finalize()

Method new()

Initialize a new instance of the class

Usage

SimulationBatchOptions$new(variableParameters = NULL, variableMolecules = NULL)

Arguments

variableParameters

Vector of absolute parameter paths to be varied in a simulation batch

variableMolecules

Vector of absolute molecule paths to be varied in a simulation batch

Returns

A new SimulationBatchOptions object.

Method print()

Print the object to the console

Usage

SimulationBatchOptions$print(...)

Arguments

...

Rest arguments.

[Package ospsuite version 10.0.70 Index]