SimulationBatchRunValues {ospsuite}	R Documentation

SimulationBatchRunValues

Description

Options to be passed to the SimulationBatch run

Super classes

ospsuite::Printable -> ospsuite::DotNetWrapper -> SimulationBatchRunValues

Active bindings

parameterValues

Vector of parameter values used in a batch run

initialValues

Vector of initial values used in a batch run

Methods

Public methods

SimulationBatchRunValues$new()

SimulationBatchRunValues$print()

Inherited methods

ospsuite::DotNetWrapper$finalize()

Method new()

Initialize a new instance of the class

Usage

SimulationBatchRunValues$new(parameterValues = NULL, initialValues = NULL)

Arguments

parameterValues

Vector of parameter values

initialValues

Vector of molecule initial values

Returns

A new SimulationBatchRunValues object.

Method print()

Print the object to the console

Usage

SimulationBatchRunValues$print(...)

Arguments

...

Rest arguments.

[Package ospsuite version 10.0.70 Index]