SimulationPKAnalyses {ospsuite}	R Documentation

SimulationPKAnalyses

Description

pK-Analyses of a simulation (either individual or population simulation).

Super classes

ospsuite::Printable -> ospsuite::DotNetWrapper -> SimulationPKAnalyses

Active bindings

simulation

Reference to the Simulation used to calculate or import the PK-Analyses (Read-Only)

allPKParameterNames

Returns the name of all pk parameters for which a value is available

allQuantityPaths

Returns the path of all quantities for which pk parameters were calculated

Methods

Public methods

SimulationPKAnalyses$new()

SimulationPKAnalyses$allPKParametersFor()

SimulationPKAnalyses$pKParameterFor()

SimulationPKAnalyses$print()

Inherited methods

ospsuite::DotNetWrapper$finalize()

Method new()

Initialize a new instance of the class

Usage

SimulationPKAnalyses$new(ref, simulation)

Arguments

ref

.NET reference

simulation

Simulation for which the pkParameters were calculated

Returns

A new SimulationPKAnalyses object.

Method allPKParametersFor()

Returns all QuantityPKParameter defined for a given path

Usage

SimulationPKAnalyses$allPKParametersFor(quantityPath)

Arguments

quantityPath

Path for which pkParameters should be retrieved

Method pKParameterFor()

The pK Parameter defined for the given path and name

Usage

SimulationPKAnalyses$pKParameterFor(quantityPath, pkParameter)

Arguments

quantityPath

Path for which the pkParameter named pkParameter should be retrieved

pkParameter

Name of the pkParameter to retrieve

Method print()

Print the object to the console

Usage

SimulationPKAnalyses$print(...)

Arguments

...

Rest arguments.

[Package ospsuite version 10.0.70 Index]