SimulationResults {ospsuite}	R Documentation

SimulationResults

Description

Results of a simulation run (either individual or population simulation)

Super classes

ospsuite::Printable -> ospsuite::DotNetWrapper -> SimulationResults

Active bindings

count

the number of individual results (Count==1 generally means that we are dealing with an individual simulation results).

simulation

Reference to the Simulation used to calculate or import the results (Read-Only).

timeValues

Vector of simulated time output values

allQuantityPaths

List of all paths for which results are defined.

allIndividualIds

List of Ids of all individuals that have been simulated

Methods

Public methods

SimulationResults$new()

SimulationResults$hasResultsForIndividual()

SimulationResults$getValuesByPath()

SimulationResults$resultsForIndividual()

SimulationResults$print()

Inherited methods

ospsuite::DotNetWrapper$finalize()

Method new()

Initialize a new instance of the class

Usage

SimulationResults$new(ref, simulation)

Arguments

ref

.NET Instance

simulation

Reference to the simulation object used to calculated the results

Returns

A new SimulationResults object.

Method hasResultsForIndividual()

Returns TRUE if results are available for the individual with id individualId otherwise FALSE

Usage

SimulationResults$hasResultsForIndividual(individualId)

Arguments

individualId

Id of the individual

Method getValuesByPath()

Returns TRUE if results are available for the individual with id individualId otherwise FALSE

Usage

SimulationResults$getValuesByPath(path, individualIds, stopIfNotFound = TRUE)

Arguments

path

Path for which values should be retrieved

individualIds

One or more individual ids for which values should be returned

stopIfNotFound

If TRUE (default) an error is thrown if no values could be found for the path/ If FALSE, a list of NA values is returned

Method resultsForIndividual()

Returns all available results for the individual with id individualId

Usage

SimulationResults$resultsForIndividual(individualId)

Arguments

individualId

Id for which the results should be returned

Method print()

Print the object to the console

Usage

SimulationResults$print(...)

Arguments

...

Rest arguments.

[Package ospsuite version 10.0.70 Index]