SimulationSettings {ospsuite}	R Documentation

SimulationSettings

Description

Settings associated with a given simulation

Super classes

ospsuite::Printable -> ospsuite::DotNetWrapper -> SimulationSettings

Active bindings

solver

Container containing all solver parameters for the simulation (read-only)

outputSelections

All selected quantities (species, observers, parameters) that will be part of the simulated results

outputSchema

OutputSchema object containing the output intervals used to generate simulation data

Methods

Public methods

SimulationSettings$print()

Inherited methods

ospsuite::DotNetWrapper$finalize()

ospsuite::DotNetWrapper$initialize()

Method print()

Print the object to the console

Usage

SimulationSettings$print(...)

Arguments

...

Rest arguments.

[Package ospsuite version 10.0.70 Index]