SnapshotParameter {ospsuite}	R Documentation

SnapshotParameter

Description

A parameter typically used in the definition of IndividualCharacteristics covariates (Height, Weight etc...)

Super classes

ospsuite::Printable -> ospsuite::DotNetWrapper -> SnapshotParameter

Active bindings

value

Parameter value

unit

Unit in which the value is defined

Methods

Public methods

SnapshotParameter$new()

SnapshotParameter$print()

SnapshotParameter$printValue()

SnapshotParameter$clone()

Inherited methods

ospsuite::DotNetWrapper$finalize()

Method new()

Initialize a new instance of the class

Usage

SnapshotParameter$new(ref = NULL, value = NULL, unit = NULL)

Arguments

ref

Optional .NET reference object. If not defined, a new instance will be created

value

Optional value of the parameter.

unit

Optional unit of the value specified.

Returns

A new SnapshotParameter object.

Method print()

Print the object to the console

Usage

SnapshotParameter$print(...)

Arguments

...

Rest arguments.

Method printValue()

Print the the parameter in one line

Usage

SnapshotParameter$printValue(caption)

Arguments

caption

Caption to display before the value of the parameter

Method clone()

The objects of this class are cloneable with this method.

Usage

SnapshotParameter$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

[Package ospsuite version 10.0.70 Index]