UserDefinedPKParameter {ospsuite}	R Documentation

UserDefinedPKParameter

Description

Definition of a user defined PKParameter that can be calculated on top of the standard PK Parameters

Super classes

ospsuite::Printable -> ospsuite::DotNetWrapper -> ospsuite::PKParameter -> UserDefinedPKParameter

Active bindings

startTime

Start time for the calculation of the PK-Parameter. If not specified, the time will start at the first time point of the simulation (optional)

startTimeOffset

Offset in min to apply to the start time (0 by default).

endTime

End time for the calculation of the PK-Parameter. If not specified, the time will end at the last time point of the simulation (optional)

endTimeOffset

Offset in min to apply to the end time (0 by default).

startApplicationIndex

1-based Index of the application to use to determine the start time for the calculation of the PK-Parameter. If not specified, the time will start at the first time point of the simulation (optional)

endApplicationIndex

1-based Index of the application to use to determine the end time for the calculation of the PK-Parameter. If not specified, the time will end at the last time point of the simulation (optional)

normalizationFactor

Factor to use to normalized the calculated PK-Parameter. (typically DrugMass, Dose, Dose per bodyweight). It is the responsibility of the caller to ensure that the value is in the correct unit. (optional)

concentrationThreshold

Used in conjunction with the threshold parameter type. If defined, the time at which this concentration was reached will be calculated

standardPKParameter

Based parameter to use to perform the PK-Analysis calculation. See StandardPKParameter enum for all possible pk parameters

Methods

Public methods

UserDefinedPKParameter$print()

UserDefinedPKParameter$clone()

Inherited methods

ospsuite::DotNetWrapper$finalize()

ospsuite::DotNetWrapper$initialize()

Method print()

Print the object to the console

Usage

UserDefinedPKParameter$print(...)

Arguments

...

Rest arguments.

Method clone()

The objects of this class are cloneable with this method.

Usage

UserDefinedPKParameter$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

[Package ospsuite version 10.0.70 Index]