.ImporterErrorTypeToDataSetErrorType {ospsuite}	R Documentation

Mapping of string representation for the error types supported by DataSet to the values supported in the importer configuration

Description

Mapping of string representation for the error types supported by DataSet to the values supported in the importer configuration

Usage

.ImporterErrorTypeToDataSetErrorType

Format

An object of class list of length 2.

[Package ospsuite version 10.0.70 Index]