Extract aggregated simulated data

Extract aggregated simulated data

Usage

.extractAggregatedSimulatedData(simData, quantiles = c(0.05, 0.5, 0.95))

Arguments

simData

A data frame with simulated data from DataCombined$toDataFrame().

quantiles

A numerical vector with quantile values (Default: c(0.05, 0.50, 0.95)), with the quantile values defining the aggregation of individual data. In the profile plot, the middle value will be used to draw a line, while the lower and upper values will be used to create a ribbon.

Details

The simulated values will be aggregated across individuals for each time point.

See also

Examples

# let's create a data frame to test this function
df <- dplyr::tibble(
  xValues = c(
    0, 1, 2, 3, 4, 5, 0, 1, 2, 3, 4, 5,
    0, 1, 2, 3, 4, 5, 0, 1, 2, 3, 4, 5, 0, 1, 2, 3, 4, 5, 0, 1, 2,
    3, 4, 5, 0, 1, 2, 3, 4, 5, 0, 1, 2, 3, 4, 5
  ),
  yValues = c(
    0,
    0.990956723690033, 0.981773018836975, 0.972471475601196, 0.963047087192535,
    0.953498184680939, 0, 0.990953505039215, 0.981729507446289, 0.97233647108078,
    0.962786376476288, 0.953093528747559, 0, 0.990955889225006, 0.981753170490265,
    0.972399413585663, 0.962896287441254, 0.953253626823425, 0, 0.990950107574463,
    0.981710314750671, 0.972296476364136, 0.962724387645721, 0.953009009361267,
    0, 0.261394888162613, 0.266657412052155, 0.27151620388031, 0.275971591472626,
    0.280027687549591, 0, 0.26139160990715, 0.266613900661469, 0.271381109952927,
    0.275710910558701, 0.279623001813889, 0, 0.261393994092941, 0.266637593507767,
    0.271443992853165, 0.275820910930634, 0.279783099889755, 0, 0.261388212442398,
    0.266594797372818, 0.27134120464325, 0.275649011135101, 0.279538512229919
  ),
  group = c(rep("Stevens 2012 solid total", 24), rep("Stevens 2012 solid distal", 24)),
  name = group
)

# raw data
df
#> # A tibble: 48 x 4
#>    xValues yValues group                    name                    
#>      <dbl>   <dbl> <chr>                    <chr>                   
#>  1       0   0     Stevens 2012 solid total Stevens 2012 solid total
#>  2       1   0.991 Stevens 2012 solid total Stevens 2012 solid total
#>  3       2   0.982 Stevens 2012 solid total Stevens 2012 solid total
#>  4       3   0.972 Stevens 2012 solid total Stevens 2012 solid total
#>  5       4   0.963 Stevens 2012 solid total Stevens 2012 solid total
#>  6       5   0.953 Stevens 2012 solid total Stevens 2012 solid total
#>  7       0   0     Stevens 2012 solid total Stevens 2012 solid total
#>  8       1   0.991 Stevens 2012 solid total Stevens 2012 solid total
#>  9       2   0.982 Stevens 2012 solid total Stevens 2012 solid total
#> 10       3   0.972 Stevens 2012 solid total Stevens 2012 solid total
#> # ... with 38 more rows

# aggregated data
ospsuite:::.extractAggregatedSimulatedData(df)
#> # A tibble: 12 x 6
#>    group                     name                xValues yValu~1 yValues yValu~2
#>    <chr>                     <chr>                 <dbl>   <dbl>   <dbl>   <dbl>
#>  1 Stevens 2012 solid distal Stevens 2012 solid~       0   0       0       0    
#>  2 Stevens 2012 solid distal Stevens 2012 solid~       1   0.261   0.261   0.261
#>  3 Stevens 2012 solid distal Stevens 2012 solid~       2   0.267   0.267   0.267
#>  4 Stevens 2012 solid distal Stevens 2012 solid~       3   0.271   0.271   0.272
#>  5 Stevens 2012 solid distal Stevens 2012 solid~       4   0.276   0.276   0.276
#>  6 Stevens 2012 solid distal Stevens 2012 solid~       5   0.280   0.280   0.280
#>  7 Stevens 2012 solid total  Stevens 2012 solid~       0   0       0       0    
#>  8 Stevens 2012 solid total  Stevens 2012 solid~       1   0.991   0.991   0.991
#>  9 Stevens 2012 solid total  Stevens 2012 solid~       2   0.982   0.982   0.982
#> 10 Stevens 2012 solid total  Stevens 2012 solid~       3   0.972   0.972   0.972
#> 11 Stevens 2012 solid total  Stevens 2012 solid~       4   0.963   0.963   0.963
#> 12 Stevens 2012 solid total  Stevens 2012 solid~       5   0.953   0.953   0.953
#> # ... with abbreviated variable names 1: yValuesLower, 2: yValuesHigher