Geometric Range

Geometric Range

Usage

.geoRange(x, nsd = 1, na.rm = FALSE, ...)

Arguments

x

numeric vector to compute geometric range from

nsd

optional argument defining the number of standard deviation to add and substract to the mean

na.rm

a logical evaluating to TRUE or FALSE indicating whether NA values should be stripped before the computation proceeds.

...

further arguments passed to mean and sd functions.

Value

numeric vector of length 3 representing the min, mean and max of the geometric range.