.getPathToPKSimInstallDirFromFileSystem {ospsuite}	R Documentation

Tries to find the installation path for a specific version of PK-Sim via the filesystem. Searching is done in the following order:

Search via filesystem in a guessed installation folder from the base.search.folder

Search via filesystem for PKSim.exe recursively from the defined base.search.folder (fallback)

Description

Tries to find the installation path for a specific version of PK-Sim via the filesystem. Searching is done in the following order:

Search via filesystem in a guessed installation folder from the base.search.folder

Search via filesystem for PKSim.exe recursively from the defined base.search.folder (fallback)

Usage

.getPathToPKSimInstallDirFromFileSystem(
  pksim.version,
  base.search.folder = Sys.getenv("ProgramW6432")
)

Arguments

pksim.version
The version number of Pk-Sim as a string.

base.search.folder
The base folder for filesystem-lookup fallback (default: 64-bit program folder)

Value

The path to the PK-Sim installation for version pksim.version or NA if no path could be found. The path is separated with slashes (unix-style) and in compliance with base-R without a trailing slash. If more than one matching path is found a warning is produced.

Examples

## Not run: 
path <- .getPathToPKSimInstallDirFromFileSystem("7.4")
path2 <- .getPathToPKSimInstallDirFromFileSystem("7.5", "C:/MyOSPFolder/")

## End(Not run)

[Package ospsuite version 10.0.70 Index]