.getPathToPKSimInstallDirFromRegistry {ospsuite}	R Documentation

Tries to find the installation path for a specific version of PK-Sim from an Windows registry entry.

Description

Tries to find the installation path for a specific version of PK-Sim from an Windows registry entry.

Usage

.getPathToPKSimInstallDirFromRegistry(pksim.version)

Arguments

pksim.version
The version number of Pk-Sim as a string.

Value

The path to the PK-Sim installation for version pksim.version or NA if no path could be found. The path is separated with slashes (unix-style) and in compilance with base-R without a trailing slash.

Examples

## Not run: 
path <- .getPathToPKSimInstallDirFromRegistry("9")

## End(Not run)

[Package ospsuite version 10.0.70 Index]