Converts a list of .NET ParameterValue into a list with 2 entries: paths, values. A 3rd optional entry units will be defined if the parameter addUnits is set to TRUE. Note: Units are only available for .NET object of type ParameterValueWithUnit

Converts a list of .NET ParameterValue into a list with 2 entries: paths, values. A 3rd optional entry units will be defined if the parameter addUnits is set to TRUE. Note: Units are only available for .NET object of type ParameterValueWithUnit

Usage

.parameterValueListFrom(netParameterValues, addUnits = FALSE)

Arguments

netParameterValues

List of.NET ParameterValue or ParameterValueWithUnit

addUnits

If TRUE, a a third list will be returned containing the units in which the parameters are defined. Default is FALSE

Value

A list with 3 sublist: paths, values, and optionally units containing the corresponding values from each parameter value