encodeUnit {ospsuite}	R Documentation

This is required to ensure that we have no issue using the mu symbol in different OS See https://github.com/Open-Systems-Pharmacology/OSPSuite-R/issues/476 for details

Description

This is required to ensure that we have no issue using the mu symbol in different OS See https://github.com/Open-Systems-Pharmacology/OSPSuite-R/issues/476 for details

Usage

encodeUnit(unit)

Arguments

unit
Unit to encode

[Package ospsuite version 10.0.70 Index]