enumGetValue {ospsuite}	R Documentation

Return the value that is stored under the given key. If the key is not present, an error is thrown.

Description

Return the value that is stored under the given key. If the key is not present, an error is thrown.

Usage

enumGetValue(enum, key)

Arguments

enum
enum that contains the key-value pair

key
key under which the value is stored

Value

Value that is assigned to key

[Package ospsuite version 10.0.70 Index]