enumPut {ospsuite}	R Documentation

Add a new key-value pairs to an enum.

Description

Add a new key-value pairs to an enum.

Usage

enumPut(keys, values, enum, overwrite = FALSE)

Arguments

keys
Keys of the values to be added

values
Values to be added

enum
enum the key-value pairs should be added to. WARNING: the original object is not modified!

overwrite
if TRUE and a value with any of the given keys exists, it will be overwritten with the new value. Otherwise, an error is thrown. Default is FALSE.

Value

Enum with added key-value pair.

Examples

myEnum <- enum(c(a = "b"))
myEnum <- enumPut("c", "d", myEnum)
myEnum <- enumPut(c("c", "d", "g"), c(12, 2, "a"), myEnum, overwrite = TRUE)

[Package ospsuite version 10.0.70 Index]