getAllEntitiesMatching {ospsuite}	R Documentation

Retrieves all entities of a container (simulation or container instance) matching the given path criteria.

Description

Retrieves all entities of a container (simulation or container instance) matching the given path criteria.

Usage

getAllEntitiesMatching(paths, container, entityType, method = NULL)

Arguments

paths
A vector of strings representing the paths relative to the container

container
A Container or Simulation used to find the entities

entityType
Class of the type that should be returned.

method
Method to call in the underlying .NET class. (optional). If unspecified, the method will be estimated from entity type

Value

A list of entities matching the path criteria coerced to the entityType. The list is empty if no entities matching were found.

See Also

loadSimulation(), getContainer() and getAllContainersMatching() to create objects of type Container or Simulation

[Package ospsuite version 10.0.70 Index]