getAllEntityPathsIn {ospsuite}	R Documentation

Retrieves all path of entities defined within the container (simulation or container instance)

Description

Retrieves all path of entities defined within the container (simulation or container instance)

Usage

getAllEntityPathsIn(container, entityType, method = NULL)

Arguments

container
A Container or Simulation used to find the entities

entityType
Type of entity for which the path should be returned.

method
Method to call in the underlying .NET class. (optional). If unspecified, the method will be estimated from entity type

Value

An array of paths (one for each entity found under the container and its sub containers) The list is empty if no entities matching were found.

See Also

loadSimulation(), getContainer() and getAllContainersMatching() to create objects of type Container or Simulation

[Package ospsuite version 10.0.70 Index]