getAllStateVariablesPaths {ospsuite}	R Documentation

Get the paths of all state variable quantities of the simulation

Description

Get the paths of all state variable quantities of the simulation

Usage

getAllStateVariablesPaths(simulation)

Arguments

simulation
Simulation object

Details

List of paths of all molecules in all compartments and all parameters that are state variables.

Value

A list of paths

[Package ospsuite version 10.0.70 Index]