getDimensionByName {ospsuite}	R Documentation

Returns the an instance of the dimension with the given name if found or NULL otherwise

Description

Returns the an instance of the dimension with the given name if found or NULL otherwise

Usage

getDimensionByName(name)

Arguments

name
Name of dimension that should be retrieved

Examples

dim <- getDimensionByName("Time")

[Package ospsuite version 10.0.70 Index]