getDimensionForUnit {ospsuite}	R Documentation

Returns the name of dimension that can be used to support the given unit or null if the dimension cannot be found

Description

Returns the name of dimension that can be used to support the given unit or null if the dimension cannot be found

Usage

getDimensionForUnit(unit)

Arguments

unit
Unit used to find the corresponding dimension

Examples

dim <- getDimensionForUnit("mg")

[Package ospsuite version 10.0.70 Index]