getEnumKey {ospsuite}	R Documentation

Get the key mapped to the given value in an enum

Description

Get the key mapped to the given value in an enum

Usage

getEnumKey(enum, value)

Arguments

enum
The enum where the key-value pair is stored

value
The value that is mapped to the key

Value

Key under which the value is stored. If the value is not in the enum, NULL is returned

[Package ospsuite version 10.0.70 Index]