getQuantity {ospsuite}	R Documentation

Retrieve a single quantity by path in the given container

Description

Retrieve a single quantity by path in the given container

Usage

getQuantity(path, container, stopIfNotFound = TRUE)

Arguments

path
A string representing the path relative to the container

container
A Container or Simulation used to find the parameters

stopIfNotFound
Boolean. If TRUE (default) and no quantity exists for the given path, an error is thrown. If FALSE, NULL is returned.

Value

The Quantity with the given path. If the quantity for the path does not exist, an error is thrown if stopIfNotFound is TRUE (default), otherwise NULL

See Also

loadSimulation(), getContainer() and getAllContainersMatching() to retrieve objects of type Container or Simulation

Examples

simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)
quantity <- getQuantity("Organism|Liver|Volume", sim)

[Package ospsuite version 10.0.70 Index]