getStandardMoleculeParameters {ospsuite}	R Documentation

Returns a list containing all standard global parameters defined in a simulation for given moleculeName. These parameters are typically located directly under the container named after the moleculeName. For the list of standard parameters

Description

Returns a list containing all standard global parameters defined in a simulation for given moleculeName. These parameters are typically located directly under the container named after the moleculeName. For the list of standard parameters

Usage

getStandardMoleculeParameters(moleculeName, simulation)

Arguments

moleculeName
Name of molecule (Enzyme, Transporter etc..) for which global parameters should be returned

simulation
Simulation to query for molecule parameters

Value

A list of all standard global parameters defined for moleculeName if the molecule exists in the simulation. Otherwise an empty list is returned

See Also

MoleculeParameter

Examples

simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim1 <- loadSimulation(simPath)

parameters <- getStandardMoleculeParameters("CYP3A4", sim1)

[Package ospsuite version 10.0.70 Index]