getUnitsForDimension {ospsuite}	R Documentation

Returns a vector containing all units defined in the dimension

Description

Returns a vector containing all units defined in the dimension

Usage

getUnitsForDimension(dimension)

Arguments

dimension
Name of dimension for which units should be returned

Examples

units <- getUnitsForDimension("Mass")

[Package ospsuite version 10.0.70 Index]