initPKSim {ospsuite}	R Documentation

Loads the PKSim.R that will enable create individual and create population workflows.

Description

Loads the PKSim.R that will enable create individual and create population workflows.

Usage

initPKSim(pksimFolderPath = NULL)

Arguments

pksimFolderPath
Path where PK-Sim is installed. If this is not specified, path will be read from registry using the package version

Note

This will only work on Windows machine and should not be called on any other OS. This function should also only be called explicitly when using a portable install of the package

[Package ospsuite version 10.0.70 Index]