isOfType {ospsuite}	R Documentation

Check if the provided object is of certain type

Description

Check if the provided object is of certain type

Usage

isOfType(object, type, nullAllowed = FALSE)

Arguments

object
An object or a list of objects

type
String representation or Class of the type that should be checked for

nullAllowed
Boolean flag if NULL is accepted for the object. If TRUE, NULL always returns TRUE, otherwise NULL returns FALSE. Default is FALSE

Value

TRUE if the object or all objects inside the list are of the given type. Only the first level of the given list is considered.

[Package ospsuite version 10.0.70 Index]