loadDataSetsFromExcel {ospsuite}	R Documentation

Load data sets from excel

Description

Load data sets from excel

Usage

loadDataSetsFromExcel(
  xlsFilePath,
  importerConfiguration,
  importAllSheets = FALSE
)

Arguments

xlsFilePath
Path to the excel file with the data

importerConfiguration
An object of type DataImporterConfiguration that is valid for the excel file

importAllSheets
If FALSE (default), only sheets specified in the importerConfiguration will be loaded. If TRUE, an attempt to load all sheets is performed. If any sheet does not comply with the configuration, an error is thrown.

Details

Load observed data from an excel file using an importer configuration

Value

A named set of DataSet objects. The naming is defined by the property importerConfiguration$namingPattern.

Examples

## Not run: 
configurationPath <- "../dataImporterConfiguration_noSheets.xml"
xlsFilePath <- "../CompiledDataSet_oneSheet.xlsx"
importerConfiguration <- DataImporterConfiguration$new(configurationPath)

dataSets <- loadDataSetsFromExcel(xlsFilePath = xlsFilePath, importerConfiguration = importerConfiguration, importAllSheets = TRUE)

## End(Not run)

[Package ospsuite version 10.0.70 Index]