loadPopulation {ospsuite}	R Documentation

Loads a population from a csv file and returns the population.

Description

Loads a population from a csv file and returns the population.

Usage

loadPopulation(csvPopulationFile)

Arguments

csvPopulationFile
Full path of csv population file to load.

Examples

csvPath <- system.file("extdata", "pop.csv", package = "ospsuite")

population <- loadPopulation(csvPath)

[Package ospsuite version 10.0.70 Index]