netEnumName {ospsuite}	R Documentation

Retrieves the name of the constant in the specified enumeration that has the specified value.

Description

Retrieves the name of the constant in the specified enumeration that has the specified value.

Usage

netEnumName(enumType, enumValue)

Arguments

enumType
a CLR object, System.Type or type name, possibly namespace and assembly qualified type name, e.g. 'My.Namespace.MyClass,MyAssemblyName'.

enumValue
The value of a particular enumerated constant in terms of its underlying type. Typically an integer.

Value

A string containing the name of the enumerated constant in enumType whose value is enumValue; or null if no such constant is found.

[Package ospsuite version 10.0.70 Index]