populationAsDataFrame {ospsuite}	R Documentation

Creates a data.frame containing one column for each parameter defined in the population

Description

Creates a data.frame containing one column for each parameter defined in the population

Usage

populationAsDataFrame(population)

Arguments

population
Population to convert to data frame (typically imported from file using loadPopulation)

Examples

csvPath <- system.file("extdata", "pop.csv", package = "ospsuite")

population <- loadPopulation(csvPath)
df <- populationAsDataFrame(population)

[Package ospsuite version 10.0.70 Index]