setMoleculeScaleDivisors {ospsuite}	R Documentation

Set molecule scale divisors

Description

Set molecule scale divisors

Usage

setMoleculeScaleDivisors(molecules, values)

Arguments

molecules
A single or a list of Molecule

values
A numeric value that should be assigned to the molecule scale factor or a vector of numeric values, if the scale factor of more than one molecule should be changed. Must have the same length as molecules

See Also

getMolecule() and getAllMoleculesMatching() to retrieve objects of type Molecule

Examples

simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)
molecule <- getMolecule("Organism|Liver|A", sim)
setMoleculeScaleDivisors(molecule, 0.001)
molecules <- getAllMoleculesMatching("Organism|**|A", sim)
setMoleculeScaleDivisors(molecules, c(0.002, 0.003))

[Package ospsuite version 10.0.70 Index]