Converts a SimulationResults objects to a data.frame

Converts a SimulationResults objects to a data.frame

Usage

simulationResultsToDataFrame(
  simulationResults,
  quantitiesOrPaths = NULL,
  population = NULL,
  individualIds = NULL
)

Arguments

simulationResults

Object of type SimulationResults produced by calling runSimulation on a Simulation object.

quantitiesOrPaths

Quantity instances (element or vector) typically retrieved using getAllQuantitiesMatching or quantity path (element or vector of strings) for which the results are to be returned. (optional) When providing the paths, only absolute full paths are supported (i.e., no matching with '*' possible). If quantitiesOrPaths is NULL (default value), returns the results for all output defined in the results.

population

population used to calculate the simulationResults (optional). This is used only to add the population covariates to the resulting data table.

individualIds

numeric IDs of individuals for which the results should be extracted. By default, all individuals from the results are considered. If the individual with the provided ID is not found, the ID is ignored.

Value

SimulationResults object as data.frame with columns IndividualId, Time, paths, simulationValues, unit, dimension, TimeUnit.

Examples

library(ospsuite)

simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
sim <- loadSimulation(simPath)

# Running an individual simulation
# results is an instance of `SimulationResults`
results <- runSimulations(sim)

# convert to a dataframe
simulationResultsToDataFrame(results)
#> # A tibble: 84 x 9
#>    paths            IndividualId  Time simulationValues unit  dimension TimeUnit
#>    <chr>                   <int> <dbl>            <dbl> <chr> <chr>     <chr>   
#>  1 Organism|Liver|A            0     0            30    µmol  Amount    min     
#>  2 Organism|Liver|A            0    15            25.8  µmol  Amount    min     
#>  3 Organism|Liver|A            0    30            22.2  µmol  Amount    min     
#>  4 Organism|Liver|A            0    45            19.1  µmol  Amount    min     
#>  5 Organism|Liver|A            0    60            16.5  µmol  Amount    min     
#>  6 Organism|Liver|A            0    75            14.2  µmol  Amount    min     
#>  7 Organism|Liver|A            0    90            12.2  µmol  Amount    min     
#>  8 Organism|Liver|A            0   105            10.5  µmol  Amount    min     
#>  9 Organism|Liver|A            0   120             9.04 µmol  Amount    min     
#> 10 Organism|Liver|A            0   135             7.79 µmol  Amount    min     
#> # ... with 74 more rows, and 2 more variables: TimeDimension <chr>,
#> #   molWeight <dbl>