toObjectType {ospsuite}	R Documentation

Transforms a single .NET object or a list of .NET Object to their corresponding wrapper class in R. Note that if the object is a single object, NULL will be returned if the .NET object is null. This allows semantic equivalence between .NET and R

Description

Transforms a single .NET object or a list of .NET Object to their corresponding wrapper class in R. Note that if the object is a single object, NULL will be returned if the .NET object is null. This allows semantic equivalence between .NET and R

Usage

toObjectType(netObject, class)

Arguments

netObject
The .NET object instances (single or list) to wrap

class
The class definition that will be used to convert the parameter

Value

The wrapped object (single or list)

[Package ospsuite version 10.0.70 Index]