toPathArray {ospsuite}	R Documentation

Convert a path defined as string to a path array

Description

Convert a path defined as string to a path array

Usage

toPathArray(path)

Arguments

path
A string representation of a path, with path entries separated by '|'

Value

An array containing one element for each path entry

Examples

array <- toPathArray("Organism|Organ|Liver")

[Package ospsuite version 10.0.70 Index]