toPathString {ospsuite}	R Documentation

Convert a path array to a path as string with entries separated by '|'

Description

Convert a path array to a path as string with entries separated by '|'

Usage

toPathString(...)

Arguments

...
Path entries to concatenate into a path string

Value

A string built using each entry of the pathArray

Examples

path <- toPathString(c("Organism", "Organ", "Liver"))

[Package ospsuite version 10.0.70 Index]