validateIsOfType {ospsuite}	R Documentation

Check if the provided object is of certain type. If not, stop with an error.

Description

Check if the provided object is of certain type. If not, stop with an error.

Usage

validateIsOfType(object, type, nullAllowed = FALSE)

Arguments

object
An object or a list of objects

type
String representation or Class of the type that should be checked for

nullAllowed
Boolean flag if NULL is accepted for the object. If TRUE, NULL is always valid, otherwise the error is thrown. Default is FALSE

[Package ospsuite version 10.0.70 Index]