DataColumn

One column defined in a DataRepository

Super classes

rSharp::NetObject -> ospsuite::DotNetWrapper -> DataColumn

Active bindings

values

Returns the values defined in the column

name

Returns the name of the column (Read-Only)

unit

The base unit in which the values are defined (Read-Only)

displayUnit

The unit in which the values should be displayed

dimension

The dimension of the values

molWeight

Molecular weight of associated observed data in internal unit In no molecular weight is defined, the value is NULL

LLOQ

Lower Limit Of Quantification. In no LLOQ is defined, the value is NULL

Methods

Public methods

DataColumn$print()

Inherited methods

rSharp::NetObject$.printClass()

rSharp::NetObject$.printLine()

rSharp::NetObject$call()

rSharp::NetObject$get()

rSharp::NetObject$getFields()

rSharp::NetObject$getMemberSignature()

rSharp::NetObject$getMethods()

rSharp::NetObject$getProperties()

rSharp::NetObject$getStaticFields()

rSharp::NetObject$getStaticMethods()

rSharp::NetObject$getStaticProperties()

rSharp::NetObject$set()

ospsuite::DotNetWrapper$initialize()

Method print()

Print the object to the console

Usage

DataColumn$print(...)

Arguments

...

Rest arguments.