DataImporterConfiguration {ospsuite}	R Documentation

DataImporterConfiguration

Description

Configuration of data import from excel or csv files. To be used with #TODO

Super classes

ospsuite::Printable -> ospsuite::DotNetWrapper -> DataImporterConfiguration

Active bindings

timeColumn

Name of the column for time values

timeUnit

If timeUnitFromColumn is FALSE, unit of the values in time column If timeUnitFromColumn is TRUE, name of the column with units of the values in time column

timeUnitFromColumn

If TRUE, units of the values in time column are defined in the column timeUnit. If FALSE, the unit is defined by timeUnit.

measurementColumn

Name of the column for measurement values

measurementDimension

If measurementUnitFromColumn is FALSE, dimension of the values in measurement column If measurementUnitFromColumn is TRUE, the dimension is guessed from the unit defined in the column measurementUnit during import process and $measurementDimension is NULL. When changing dimension, the unit is set to the base unit of this dimension.

measurementUnit

If measurementUnitFromColumn is FALSE, unit of the values in measurement column If measurementUnitFromColumn is TRUE, name of the column with units of the values in measurement column

measurementUnitFromColumn

If TRUE, units of the values in measurement column are defined in the column measurementUnit. If FALSE, the unit is defined by measurementUnit.

errorColumn

Name of the column for measurement error values If no error column is defined, the value is NULL. Setting the value to NULL removes an existing error column.

errorUnit

If measurementUnitFromColumn is FALSE, unit of the values in the error column If measurementUnitFromColumn is TRUE, name of the column with units of the values in error column If no error column is present, the value is NULL

errorType

Type of the measurement error values. See enum DataErrorType for possible values If no error column is present, the value is NULL

groupingColumns

Column names by which the data will be grouped

sheets

Names of the sheets (list of strings) of the excel workbook for which the configuration will be applied.

namingPattern

Regular expression used for naming of loaded data sets. Words between curly brackets (e.g. {Group Id}) will be replaced by the value in the corresponding column. Further keywords are {Source} for the file name and {Sheet} for sheet name.

Methods

Public methods

DataImporterConfiguration$new()

DataImporterConfiguration$saveConfiguration()

DataImporterConfiguration$addGroupingColumn()

DataImporterConfiguration$removeGroupingColumn()

DataImporterConfiguration$print()

DataImporterConfiguration$clone()

Inherited methods

ospsuite::DotNetWrapper$finalize()

Method new()

Initialize a new instance of the class

Usage

DataImporterConfiguration$new(configurationFilePath = NULL)

Arguments

configurationFilePath

Path to the XML file with stored configuration (e.g. create in PK-Sim or MoBi). If NULL (default), an empty configuration with columns "Time" and "Measurement" is created.

Returns

A new DataImporterConfiguration object.

Method saveConfiguration()

Save configuration to a XML file that can be used in PKSim/MoBi

Usage

DataImporterConfiguration$saveConfiguration(filePath)

Arguments

filePath

Path (incl. file name) to the location where the configuration will be exported to.

Method addGroupingColumn()

Add a column for grouping the data sets

Usage

DataImporterConfiguration$addGroupingColumn(column)

Arguments

column

Name of the column

Method removeGroupingColumn()

Remove a column for grouping the data sets

Usage

DataImporterConfiguration$removeGroupingColumn(column)

Arguments

column

Name of the column

Method print()

Print the object to the console

Usage

DataImporterConfiguration$print(...)

Arguments

...

Rest arguments.

Method clone()

The objects of this class are cloneable with this method.

Usage

DataImporterConfiguration$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

[Package ospsuite version 10.0.70 Index]