DataImporterConfiguration

Configuration of data import from excel or csv files. To be used with #TODO

Super classes

rSharp::NetObject -> ospsuite::DotNetWrapper -> DataImporterConfiguration

Active bindings

timeColumn

Name of the column for time values

timeUnit

If isTimeUnitFromColumn is FALSE, unit of the values in time column If isTimeUnitFromColumn is TRUE, name of the column with units of the values in time column.

isTimeUnitFromColumn

If TRUE, units of the values in time column are defined in the column timeUnit. If FALSE, the unit is defined by the value of timeUnit.

measurementColumn

Name of the column for measurement values

lloqColumn

Name of the column for LLOQ values If the column name is not set (value NULL), LLOQ values will be imported from the measurement column if values are written in the form '< xxx' (e.g., '<0.001'). Otherwise, the values will be imported from the specified column

measurementDimension

If isMeasurementUnitFromColumn is FALSE, dimension of the values in measurement column If isMeasurementUnitFromColumn is TRUE, the dimension is guessed from the unit defined in the column measurementUnit during import process and $measurementDimension is NULL. When changing dimension, the unit is set to the base unit of this dimension.

measurementUnit

If isMeasurementUnitFromColumn is FALSE, unit of the values in measurement column If isMeasurementUnitFromColumn is TRUE, name of the column with units of the values in measurement column

isMeasurementUnitFromColumn

If TRUE, units of the values in measurement column are defined in the column measurementUnit. If FALSE, the unit is defined by the value of measurementUnit.

errorColumn

Name of the column for measurement error values If no error column is defined, the value is NULL. Setting the value to NULL removes an existing error column.

errorUnit

If isMeasurementUnitFromColumn is FALSE, unit of the values in the error column If isMeasurementUnitFromColumn is TRUE, name of the column with units of the values in error column If no error column is present, the value is NULL

errorType

Type of the measurement error values. See enum DataErrorType for possible values If no error column is present, the value is NULL

groupingColumns

Column names by which the data will be grouped

sheets

Names of the sheets (list of strings) of the excel workbook for which the configuration will be applied.

namingPattern

Regular expression used for naming of loaded data sets. Words between curly brackets (e.g. {Group Id}) will be replaced by the value in the corresponding column. Further keywords are {Source} for the file name and {Sheet} for sheet name.

Methods

Public methods

DataImporterConfiguration$new()

DataImporterConfiguration$saveConfiguration()

DataImporterConfiguration$addGroupingColumn()

DataImporterConfiguration$removeGroupingColumn()

DataImporterConfiguration$print()

Inherited methods

rSharp::NetObject$.printClass()

rSharp::NetObject$.printLine()

rSharp::NetObject$call()

rSharp::NetObject$get()

rSharp::NetObject$getFields()

rSharp::NetObject$getMemberSignature()

rSharp::NetObject$getMethods()

rSharp::NetObject$getProperties()

rSharp::NetObject$getStaticFields()

rSharp::NetObject$getStaticMethods()

rSharp::NetObject$getStaticProperties()

rSharp::NetObject$set()

Method new()

Initialize a new instance of the class

Usage

DataImporterConfiguration$new(netObject = NULL)

Arguments

netObject

A NetObject with the reference to .NET DataImporterConfiguration object If NULL (default), an empty configuration with columns "Time" and "Measurement" is created.

Returns

A new DataImporterConfiguration object.

Method saveConfiguration()

Save configuration to a XML file that can be used in PK-Sim/MoBi

Usage

DataImporterConfiguration$saveConfiguration(filePath)

Arguments

filePath

Path (incl. file name) to the location where the configuration will be exported to.

Method addGroupingColumn()

Add a column for grouping the data sets

Usage

DataImporterConfiguration$addGroupingColumn(column)

Arguments

column

Name of the column

Method removeGroupingColumn()

Remove a column for grouping the data sets

Usage

DataImporterConfiguration$removeGroupingColumn(column)

Arguments

column

Name of the column

Method print()

Print the object to the console

Usage

DataImporterConfiguration$print(...)

Arguments

...

Rest arguments.