IndividualCharacteristics

Characteristics of an individual describing its origin

Super classes

rSharp::NetObject -> ospsuite::DotNetWrapper -> IndividualCharacteristics

Active bindings

species

Specifies the species of the individual. It should be a species available in PK-Sim (see Species)

population

For a Human species, the population of interest. It should be a population available in PK-Sim (see HumanPopulation)

gender

Gender of the individual. It should be defined for the species in PK-Sim (see Gender)

age

Age of the individual as in instance of a SnapshotParameter (optional)

gestationalAge

Gestational Age of the individual as in instance of a SnapshotParameter (optional)

weight

Weight of the individual as in instance of a SnapshotParameter (optional)

height

Height of the individual as in instance of a SnapshotParameter (optional)

allMoleculeOntogenies

All molecule ontogenies defined for this individual characteristics.

seed

Seed used to generate the population

Methods

Public methods

IndividualCharacteristics$new()

IndividualCharacteristics$print()

IndividualCharacteristics$addMoleculeOntogeny()

Inherited methods

rSharp::NetObject$.printClass()

rSharp::NetObject$.printLine()

rSharp::NetObject$call()

rSharp::NetObject$get()

rSharp::NetObject$getFields()

rSharp::NetObject$getMemberSignature()

rSharp::NetObject$getMethods()

rSharp::NetObject$getProperties()

rSharp::NetObject$getStaticFields()

rSharp::NetObject$getStaticMethods()

rSharp::NetObject$getStaticProperties()

rSharp::NetObject$set()

Method new()

Initialize a new instance of the class

Usage

IndividualCharacteristics$new()

Returns

A new IndividualCharacteristics object.

Method print()

Print the object to the console

Usage

IndividualCharacteristics$print(...)

Arguments

...

Rest arguments.

Method addMoleculeOntogeny()

Add a molecule ontogeny MoleculeOntogeny to the individual characteristics

Usage

IndividualCharacteristics$addMoleculeOntogeny(moleculeOntogeny)

Arguments

moleculeOntogeny

Molecule ontogeny to add