Molecule {ospsuite}	R Documentation

Molecule

Description

A molecule defined in a compartment of the system

Details

Derived from Quantity, please see base class documentation.

Super classes

ospsuite::Printable -> ospsuite::DotNetWrapper -> ospsuite::ObjectBase -> ospsuite::Entity -> ospsuite::Quantity -> Molecule

Active bindings

value

Start value of the molecule

scaleDivisor

Scale divisor. Its purpose is to reduce numerical noise and to enhance computation performance. see https://docs.open-systems-pharmacology.org/working-with-mobi/mobi-documentation/model-building-components#import-molecule-and-parameter-start-values-from-excel

Methods

Public methods

Molecule$new()

Molecule$print()

Inherited methods

ospsuite::DotNetWrapper$finalize()

ospsuite::Quantity$hasUnit()

ospsuite::Quantity$printQuantityValue()

ospsuite::Quantity$printValue()

ospsuite::Quantity$reset()

ospsuite::Quantity$setValue()

Method new()

Initialize a new instance of the class

Usage

Molecule$new(ref)

Arguments

ref

.NET reference object.

Returns

A new Molecule object.

Method print()

Print the object to the console

Usage

Molecule$print(...)

Arguments

...

Rest arguments.

[Package ospsuite version 10.0.70 Index]