Molecule

A molecule defined in a compartment of the system

Details

Derived from Quantity, please see base class documentation.

Super classes

rSharp::NetObject -> ospsuite::DotNetWrapper -> ospsuite::ObjectBase -> ospsuite::Entity -> ospsuite::Quantity -> Molecule

Active bindings

value

Start value of the molecule

scaleDivisor

Scale divisor. Its purpose is to reduce numerical noise and to enhance computation performance. see https://docs.open-systems-pharmacology.org/working-with-mobi/mobi-documentation/model-building-components#import-molecule-and-parameter-start-values-from-excel

Methods

Public methods

Molecule$new()

Molecule$print()

Inherited methods

rSharp::NetObject$call()

rSharp::NetObject$finalize()

rSharp::NetObject$get()

rSharp::NetObject$getFields()

rSharp::NetObject$getMemberSignature()

rSharp::NetObject$getMethods()

rSharp::NetObject$getProperties()

rSharp::NetObject$getStaticFields()

rSharp::NetObject$getStaticMethods()

rSharp::NetObject$getStaticProperties()

rSharp::NetObject$set()

ospsuite::Quantity$hasUnit()

ospsuite::Quantity$printQuantityValue()

ospsuite::Quantity$printValue()

ospsuite::Quantity$reset()

ospsuite::Quantity$setValue()

Method new()

Initialize a new instance of the class

Usage

Molecule$new(netObject)

Arguments

netObject

An rSharp::NetObject object.

Returns

A new Molecule object.

Method print()

Print the object to the console

Usage

Molecule$print(...)

Arguments

...

Rest arguments.