OutputSchema

Output schema associated with a given simulation

Super classes

rSharp::NetObject -> ospsuite::DotNetWrapper -> OutputSchema

Active bindings

intervals

All intervals defined in the schema (Read-Only)

timePoints

All single time points defined in the schema (Read-Only)

endTime

Returns the end time of the simulation in kernel unit (Read-Only)

Methods

Public methods

OutputSchema$clear()

OutputSchema$addInterval()

OutputSchema$removeInterval()

OutputSchema$addTimePoints()

OutputSchema$print()

Inherited methods

rSharp::NetObject$.printClass()

rSharp::NetObject$.printLine()

rSharp::NetObject$call()

rSharp::NetObject$get()

rSharp::NetObject$getFields()

rSharp::NetObject$getMemberSignature()

rSharp::NetObject$getMethods()

rSharp::NetObject$getProperties()

rSharp::NetObject$getStaticFields()

rSharp::NetObject$getStaticMethods()

rSharp::NetObject$getStaticProperties()

rSharp::NetObject$set()

ospsuite::DotNetWrapper$initialize()

Method clear()

Clears all intervals and time points

Usage

OutputSchema$clear()

Method addInterval()

Adds an interval to the schema

Usage

OutputSchema$addInterval(interval)

Arguments

interval

Interval to add

Method removeInterval()

Removes the interval from the schema

Usage

OutputSchema$removeInterval(interval)

Arguments

interval

Interval to remove

Method addTimePoints()

Adds the time points to the schema. Note that time points and intervals exists concurrently. Use time points only if you need to ensure that specific time are used.

Usage

OutputSchema$addTimePoints(timePoints)

Arguments

timePoints

Time points to add to the schema

Method print()

Print the object to the console

Usage

OutputSchema$print(...)

Arguments

...

Rest arguments.