ParameterRange

A parameter range typically used in the definition of PopulationCharacteristics covariates (Height, Weight etc...)

Super classes

rSharp::NetObject -> ospsuite::DotNetWrapper -> ParameterRange

Active bindings

min

Minimum value for the parameter range

max

Maximum value for the parameter range

unit

Unit in which the value is defined

Methods

Public methods

ParameterRange$new()

ParameterRange$print()

ParameterRange$printValue()

ParameterRange$getPrintValue()

Inherited methods

rSharp::NetObject$.printClass()

rSharp::NetObject$.printLine()

rSharp::NetObject$call()

rSharp::NetObject$get()

rSharp::NetObject$getFields()

rSharp::NetObject$getMemberSignature()

rSharp::NetObject$getMethods()

rSharp::NetObject$getProperties()

rSharp::NetObject$getStaticFields()

rSharp::NetObject$getStaticMethods()

rSharp::NetObject$getStaticProperties()

rSharp::NetObject$set()

Method new()

Initialize a new instance of the class

Usage

ParameterRange$new(netObject = NULL, min = NULL, max = NULL, unit = NULL)

Arguments

netObject

Optional NetObject of ParameterRange. If not defined, a new instance will be created

min

Optional minimum value for the range

max

Optional minimum value for the range

unit

Optional unit of the specified min and max

Returns

A new ParameterRange object.

Method print()

Print the object to the console

Usage

ParameterRange$print(...)

Arguments

...

Rest arguments.

Method printValue()

Print the parameter in one line

Usage

ParameterRange$printValue(caption)

Arguments

caption

Caption to display before the value of the parameter

Method getPrintValue()

Return a string for printing the parameter in one line

Usage

ParameterRange$getPrintValue()

Returns

A string for printing the parameter in one line