Population

List of individuals used in a population simulation

Super classes

ospsuite.utils::Printable -> ospsuite::DotNetWrapper -> Population

Active bindings

count

the number of individual in the population

allCovariateNames

the names of all covariates defined in the population

allParameterPaths

the paths of all parameters defined in the population

allIndividualIds

Ids of individuals defined in the population

Methods

Public methods

Population$has()

Population$setParameterValues()

Population$getParameterValues()

Population$getCovariateValues()

Population$getCovariateValue()

Population$getParameterValuesForIndividual()

Population$remove()

Population$print()

Inherited methods

ospsuite::DotNetWrapper$initialize()

Method has()

Returns TRUE if the population has variability defined for parameterOrPath otherwise FALSE

Usage

Population$has(parameterOrPath)

Arguments

parameterOrPath

Parameter instance of parameter path

Method setParameterValues()

Updates or adds the variability values in the population for parameterOrPath.

Usage

Population$setParameterValues(parameterOrPath, values)

Arguments

parameterOrPath

Parameter instance of parameter path. If an entry already exists for this parameter by path, its values be overwritten, otherwise it will be created.

values

double vector containing the value to set for the parameterOrPath

Method getParameterValues()

Returns the variability values defined in the population for parameterOrPath

Usage

Population$getParameterValues(parameterOrPath)

Arguments

parameterOrPath

Parameter instance of parameter path

Method getCovariateValues()

Returns the values defined in the population for the covariate named covariateName

Usage

Population$getCovariateValues(covariateName)

Arguments

covariateName

Name of covariate for which values should be retrieved

Method getCovariateValue()

Returns the values defined in the population for the covariate named covariateName and individual with id individualId

Usage

Population$getCovariateValue(covariateName, individualId)

Arguments

covariateName

Name of covariate for which values should be retrieved

individualId

Id of individual for which the value for covariate covariateName should be retrieved

Method getParameterValuesForIndividual()

Returns all values defined in the population the individual with id individualId

Usage

Population$getParameterValuesForIndividual(individualId)

Arguments

individualId

Id of individual for which all values should be returned

Method remove()

Removes the value of a parameter by path

Usage

Population$remove(parameterPath)

Arguments

parameterPath

Path of the parameter values to remove

Method print()

Print the object to the console

Usage

Population$print(...)

Arguments

...

Rest arguments.