Quantity

A quantity of the model (with unit, value) such as a Parameter or an Amount

Super classes

ospsuite.utils::Printable -> ospsuite::DotNetWrapper -> ospsuite::ObjectBase -> ospsuite::Entity -> Quantity

Active bindings

value

The value of the quantity in unit

unit

The base unit in which the quantity value is defined (Read-Only)

displayUnit

The unit in which the quantity value is usually displayed (Read-Only)

dimension

The dimension in which the quantity is defined (Read-Only)

allUnits

the list of all supported units (Read-Only)

quantityType

The type of the quantity (Read-Only)

formula

An instance of a Formula object used by this quantity (Read-Only)

isTable

Returns TRUE if the formula used by this quantity is a table formula otherwise FALSE

isConstant

Returns TRUE if the formula used by this quantity is a constant formula otherwise FALSE

isFormula

Returns TRUE if the formula used by this quantity is an explicit formula (e.g an equation) otherwise FALSE

isDistributed

Returns TRUE if the quantity represents a quantity with an underlying distribution otherwise FALSE

formulaString

Returns the equation of the formula for a quantity using an explicit formula (e.g. isFormula == TRUE) or NULL for a quantity that does not use an explicit formula.

isFixedValue

Returns TRUE if the formula was overridden by a constant value, otherwise FALSE

Methods

Public methods

Quantity$new()

Quantity$print()

Quantity$printValue()

Quantity$printQuantityValue()

Quantity$setValue()

Quantity$hasUnit()

Quantity$reset()

Method new()

Initialize a new instance of the class

Usage

Quantity$new(ref)

Arguments

ref

.NET Instance

Returns

A new Quantity object.

Method print()

Print the object to the console

Usage

Quantity$print(...)

Arguments

...

Rest arguments.

Method printValue()

Print the name of the quantity and its value

Usage

Quantity$printValue()

Method printQuantityValue()

Print the the value and unit of the quantity

Usage

Quantity$printQuantityValue(caption)

Arguments

caption

Text to prepend to the value

Method setValue()

Convert value from unit to the base unit and sets the value in base unit.

Usage

Quantity$setValue(value, unit = NULL)

Arguments

value

Value to set. If unit is null, we assume that the value is in base unit

unit

Optional unit in which the value is given.

Method hasUnit()

Returns TRUE if the quantity supports the given unit otherwise FALSE. For the list of supported units, use allUnits

Usage

Quantity$hasUnit(unit)

Arguments

unit

Unit to check

Method reset()

Ensures that the quantity uses the value computed by its formula. It is a shortcut for self$isFixedValue <- false.

Usage

Quantity$reset()