SensitivityAnalysis

Supports Sensitivity Analysis workflow to assess the impact of input parameters on the simulation outputs

Super classes

rSharp::NetObject -> ospsuite::DotNetWrapper -> SensitivityAnalysis

Active bindings

simulation

Reference to the Simulation used to calculate or import the sensitivity analysis results (Read-Only).

numberOfSteps

Number of steps used for the variation of each parameter (optional, default specified in ospsuiteEnv$sensitivityAnalysisConfig)

variationRange

Variation applied to the parameter (optional, default specified in ospsuiteEnv$sensitivityAnalysisConfig)

parameterPaths

List of parameters to use for sensitivity calculation.If empty, the sensitivity will be performed automatically on all constant parameters that are really in use in the simulation. Constant parameter means all parameters with a constant value or a formula parameter with a value that was overridden by the user

Methods

Public methods

SensitivityAnalysis$new()

SensitivityAnalysis$addParameterPaths()

SensitivityAnalysis$clearParameterPaths()

SensitivityAnalysis$print()

Inherited methods

rSharp::NetObject$call()

rSharp::NetObject$finalize()

rSharp::NetObject$get()

rSharp::NetObject$getFields()

rSharp::NetObject$getMemberSignature()

rSharp::NetObject$getMethods()

rSharp::NetObject$getProperties()

rSharp::NetObject$getStaticFields()

rSharp::NetObject$getStaticMethods()

rSharp::NetObject$getStaticProperties()

rSharp::NetObject$set()

Method new()

Initialize a new instance of the class

Usage

SensitivityAnalysis$new(
  simulation,
  parameterPaths = NULL,
  numberOfSteps = ospsuiteEnv$sensitivityAnalysisConfig$numberOfSteps,
  variationRange = ospsuiteEnv$sensitivityAnalysisConfig$variationRange
)

Arguments

simulation

Simulation for which a sensitivity analysis should be performed

parameterPaths

Vector of parameter paths to use for sensitivity calculation (optional).If undefined, the sensitivity will be performed automatically on all constant parameters of the simulation. Constant parameter means all parameters with a constant value or a formula parameter with a value that was overridden by the user

numberOfSteps

Number of steps used for the variation of each parameter (optional, default specified in getOSPSuiteSetting("sensitivityAnalysisConfig"))

variationRange

Variation applied to the parameter (optional, default specified in getOSPSuiteSetting("sensitivityAnalysisConfig"))

Returns

A new SensitivityAnalysis object.

Method addParameterPaths()

Adds the parameterPaths to the list of parameter path to vary in the sensitivity analysis

Usage

SensitivityAnalysis$addParameterPaths(parameterPaths)

Arguments

parameterPaths

Parameter paths to add (single or multiple values) If no parameters were specified during creating of a SensitivityAnalysis (all constant parameters are considered), calling addParameterPaths will make only the manually added parameters being varied.

Method clearParameterPaths()

Removes all parameter paths defined in the Sensitivity Analysis

Usage

SensitivityAnalysis$clearParameterPaths()

Method print()

Print the object to the console

Usage

SensitivityAnalysis$print(...)

Arguments

...

Rest arguments.