SensitivityAnalysisResults {ospsuite}	R Documentation

SensitivityAnalysisResults

Description

Results of a sensitivity analysis run (either individual or population simulation)

Super classes

ospsuite::Printable -> ospsuite::DotNetWrapper -> SensitivityAnalysisResults

Active bindings

simulation

Reference to the Simulation used to calculate or import the sensitivity analysis results (Read-Only).

count

the number of pk parameter sensitivity entries

allPKParameterNames

Returns the name of all PK-Parameters available in this results. This will be a subset of all potential PK-Parameters available in the system

allQuantityPaths

Returns the path of all outputs available in this results.

Methods

Public methods

SensitivityAnalysisResults$new()

SensitivityAnalysisResults$allPKParameterSensitivitiesFor()

SensitivityAnalysisResults$pkParameterSensitivityValueFor()

SensitivityAnalysisResults$print()

Inherited methods

ospsuite::DotNetWrapper$finalize()

Method new()

Initialize a new instance of the class

Usage

SensitivityAnalysisResults$new(ref, simulation)

Arguments

ref

.NET Instance

simulation

Reference to the simulation object used to calculated the results

Returns

A new SensitivityAnalysisResults object.

Method allPKParameterSensitivitiesFor()

Returns the PKParameterSensitivity for a given pkParameter and output participating to a total sensitivity greater or equal to totalSensitivityThreshold

Usage

SensitivityAnalysisResults$allPKParameterSensitivitiesFor(
  pkParameterName,
  outputPath,
 
    totalSensitivityThreshold = ospsuiteEnv$sensitivityAnalysisConfig$totalSensitivityThreshold
)

Arguments

pkParameterName

Name of pkParameter for which sensitivity should be retrieved

outputPath

Path of the output for which the sensitivity should be retrieved

totalSensitivityThreshold

Threshold used to filter out the most sensitive parameter. A threshold of 0.9 means that only parameter participating to a total of 90 percent of the sensitivity would be returned. A value of 1 would return the sensitivity for all parameters.

Method pkParameterSensitivityValueFor()

Returns the sensitivity value for a given pkParameter, output and parameter. If the sensitivity result does not exist, returns NaN

Usage

SensitivityAnalysisResults$pkParameterSensitivityValueFor(
  pkParameterName,
  outputPath,
  parameterName
)

Arguments

pkParameterName

Name of pkParameter for which sensitivity should be retrieved

outputPath

Path of the output for which the sensitivity should be retrieved

parameterName

Name of the sensitivity parameter for which the sensitivity should be retrieved

Method print()

Print the object to the console

Usage

SensitivityAnalysisResults$print(...)

Arguments

...

Rest arguments.

[Package ospsuite version 10.0.70 Index]