Simulation

An OSPSuite simulation

Super classes

ospsuite.utils::Printable -> ospsuite::DotNetWrapper -> ospsuite::ObjectBase -> Simulation

Active bindings

root

Root container of the simulation (read-only)

path

Path of the root container of the simulation (read-only)

solver

SimulationSolver object for the simulation (read-only)

outputSchema

outputSchema object for the simulation (read-only)

outputSelections

outputSelections object for the simulation (read-only)

sourceFile

Path to the file the simulation was loaded from (read-only)

Methods

Public methods

Simulation$new()

Simulation$allEndogenousStationaryMoleculeNames()

Simulation$allXenobioticFloatingMoleculeNames()

Simulation$allStationaryMoleculeNames()

Simulation$allFloatingMoleculeNames()

Simulation$molWeightFor()

Simulation$allApplicationsFor()

Simulation$print()

Inherited methods

Method new()

Initialize a new instance of the class

Usage

Simulation$new(ref, sourceFile = NULL)

Arguments

ref

Reference to .NET simulation object

sourceFile

(Optional) File used to load the simulation

Returns

A new Simulation object.

Method allEndogenousStationaryMoleculeNames()

Returns the name of all endogenous stationary molecules defined in the simulation. (e.g. with the flag IsStationary = TRUE) This is a typically a molecule that is individual specific such as en Enzyme, Protein, Transporter, FcRn etc.

Usage

Simulation$allEndogenousStationaryMoleculeNames()

Method allXenobioticFloatingMoleculeNames()

Returns the name of all xenobiotoc floating molecules defined in the simulation. (e.g. with the flag IsStationary = FALSE) This is typically a molecule that is being explicitly simulated such as Compound, Inhibitor, DrugComplex.

Usage

Simulation$allXenobioticFloatingMoleculeNames()

Method allStationaryMoleculeNames()

Returns the name of all stationary molecules defined in the simulation. (e.g. with the flag IsStationary = TRUE)

Usage

Simulation$allStationaryMoleculeNames()

Method allFloatingMoleculeNames()

Returns the name of all floating molecules defined in the simulation. (e.g. with the flag IsStationary = FALSE)

Usage

Simulation$allFloatingMoleculeNames()

Method molWeightFor()

Returns the mol weight value (in core unit) associated to the quantity with given path or NA if not found

Usage

Simulation$molWeightFor(quantityPath)

Arguments

quantityPath

Path of quantity used to retrieve the molecular weight

Method allApplicationsFor()

Returns the applications ordered by start time associated to the quantity with path quantityPath or an empty list if not found

Usage

Simulation$allApplicationsFor(quantityPath)

Arguments

quantityPath

Path of quantity used to retrieve the applications (e.g. applications resulting in this quantity being applied)

Method print()

Print the object to the console

Usage

Simulation$print(...)

Arguments

...

Rest arguments.