SimulationRunOptions {ospsuite}	R Documentation

SimulationRunOptions

Description

Options to be passed to the simulation engine

Super classes

ospsuite::Printable -> ospsuite::DotNetWrapper -> SimulationRunOptions

Active bindings

numberOfCores

(Maximal) number of cores to be used. This is only relevant when simulating a population simulation. Default is getOSPSuiteSetting("numberOfCores").

checkForNegativeValues

Specifies whether negative values check is on or off. Default is TRUE

showProgress

Specifies whether progress bar should be shown during simulation run. Default is getOSPSuiteSetting("showProgress")

Methods

Public methods

SimulationRunOptions$new()

SimulationRunOptions$print()

Inherited methods

ospsuite::DotNetWrapper$finalize()

Method new()

Initialize a new instance of the class

Usage

SimulationRunOptions$new(
  numberOfCores = NULL,
  checkForNegativeValues = NULL,
  showProgress = NULL
)

Arguments

numberOfCores

Number of cores to use for the simulation. Default value is getOSPSuiteSetting("numberOfCores")

checkForNegativeValues

Should the solver check for negative values. Default is TRUE

showProgress

Should a progress information be displayed. Default value is getOSPSuiteSetting("showProgress")

Returns

A new SimulationRunOptions object.

Method print()

Print the object to the console

Usage

SimulationRunOptions$print(...)

Arguments

...

Rest arguments.

[Package ospsuite version 10.0.70 Index]