SolverSettings {ospsuite}	R Documentation

SolverSettings

Description

Solver settings associated with a given simulation

Super classes

ospsuite::Printable -> ospsuite::DotNetWrapper -> SolverSettings

Active bindings

useJacobian

Use of Jacobian matrix during calculations

h0

Initial time step size

hMin

Minimum absolute value of step size allowed

hMax

Maximum absolute value of step size allowed

mxStep

Maximum number of internal steps to be taken by the solver in its attempt to reach tout

relTol

Relative tolerance of unknowns

absTol

Absolute tolerance of unknowns

Methods

Public methods

SolverSettings$print()

Inherited methods

ospsuite::DotNetWrapper$finalize()

ospsuite::DotNetWrapper$initialize()

Method print()

Print the object to the console

Usage

SolverSettings$print(...)

Arguments

...

Rest arguments.

[Package ospsuite version 10.0.70 Index]