UserDefinedPKParameter

Definition of a user defined PKParameter that can be calculated on top of the standard PK Parameters

Super classes

rSharp::NetObject -> ospsuite::DotNetWrapper -> ospsuite::PKParameter -> UserDefinedPKParameter

Active bindings

startTime

Start time in minutes for the calculation of the PK-Parameter. If not specified, the time will start at the first time point of the simulation (optional)

startTimeOffset

Offset in minutes to apply to the start time or to the start time of the application identified by startApplicationIndex. (0 by default).

endTime

End time in minutes for the calculation of the PK-Parameter. If not specified, the time will end at the last time point of the simulation (optional)

endTimeOffset

Offset in minutes to apply to the end time or to the start time of the application identified by endApplicationIndex. (0 by default).

startApplicationIndex

1-based Index of the application to use to determine the start time for the calculation of the PK-Parameter. If not specified, the time will start at the first time point of the simulation (optional)

endApplicationIndex

1-based Index of the application to use to determine the end time for the calculation of the PK-Parameter. If not specified, the time will end at the last time point of the simulation (optional)

normalizationFactor

Factor to use to normalized the calculated PK-Parameter. (typically DrugMass, Dose, DosePerBodyWeight). It is the responsibility of the caller to ensure that the value is in the correct unit. (optional)

concentrationThreshold

Used in conjunction with the threshold parameter type. If defined, the time at which this concentration was reached will be calculated

standardPKParameter

Based parameter to use to perform the PK-Analysis calculation. See StandardPKParameter enum for all possible pk parameters

Methods

Public methods

UserDefinedPKParameter$print()

Inherited methods

rSharp::NetObject$.printClass()

rSharp::NetObject$.printLine()

rSharp::NetObject$call()

rSharp::NetObject$get()

rSharp::NetObject$getFields()

rSharp::NetObject$getMemberSignature()

rSharp::NetObject$getMethods()

rSharp::NetObject$getProperties()

rSharp::NetObject$getStaticFields()

rSharp::NetObject$getStaticMethods()

rSharp::NetObject$getStaticProperties()

rSharp::NetObject$set()

ospsuite::DotNetWrapper$initialize()

Method print()

Print the object to the console

Usage

UserDefinedPKParameter$print(...)

Arguments

...

Rest arguments.