Get the paths of all state variable quantities of the simulation

Get the paths of all state variable quantities of the simulation

Usage

getAllStateVariablesPaths(simulation)

Arguments

simulation

Simulation object

Value

A list of paths

Details

List of paths of all molecules in all compartments and all parameters that are state variables.